RFC: include timezone info in all dagster log timestamps
Summary:
Before I go in and fix the large number of tests this is likely to break, I wanted to see if this is something we even want to do. Adding timezone info to logs (both the dagit/daemon logs, but also run logs that are logged oto console)seems like a useful thing to me.

Concretely, before this change console output during a run might look something like:

`
2021-12-07 23:22:17 - dagster - DEBUG - asset_lineage_pipeline - 8b862d6a-4a75-424e-a52f-08455e7f5b37 - 6145 - daily_top_action - STEP_SUCCESS - Finished execution of step "daily_top_action" in 104ms.
`

Now, it will look like:
`
2021-12-07 23:22:17-0600 - dagster - DEBUG - asset_lineage_pipeline - 8b862d6a-4a75-424e-a52f-08455e7f5b37 - 6145 - daily_top_action - STEP_SUCCESS - Finished execution of step "daily_top_action" in 104ms.
`